RavenDB-18353 SortingMatch cleanup + alphanumeric test
